Extract checked_interpolate_impl! macro for Interp1D/2D/3D
interpolate and interpolate_fast's trait-forwarding bodies (convert
the incoming slice to [D::Elem; N], then call the type's own inherent
method) were identical across Interp1D/2D/3D modulo N, mirroring the
existing extrapolate_impl!/partialeq_impl! pattern for this kind of
per-dimensionality boilerplate. InterpND has no fixed N and keeps its
own hand-written implementations of both methods.
